requirements.txt
langchain
langchain-openai
langchain-community
streamlit
faiss-cpu
tiktoken
python-dotenv
nltk
langgraph
streamlit-mermaid
google-api-python-client
google-auth-oauthlib
